+++ src/builtins/walk_sensors.rs
@@ -1,0 +1,21 @@
+use crate::sensor_sets::TnuaSensors;
+use bevy::prelude::*;
+use bevy_tnua_physics_integration_layer::data_for_backends::TnuaProximitySensor;
+
+#[derive(Copy, Clone)]
+pub struct TnuaBuiltinWalkSensors<'a> {
+    pub ground: &'a TnuaProximitySensor,
+    // TODO:
+    //pub overhead: Option<&'a TnuaProximitySensor>,
+}
+
+impl<'a> TnuaSensors<'a> for TnuaBuiltinWalkSensors<'a> {
+    type Entities = TnuaBuiltinWalkSensorsEntities;
+}
+
+#[derive(Default)]
+pub struct TnuaBuiltinWalkSensorsEntities {
+    pub ground: Option<Entity>,
+    // TODO:
+    //pub overhead: Option<Entity>,
+}