+++ src/lib.rs
@@ -134,6 +134,7 @@ pub mod builtins;
 pub mod control_helpers;
 pub mod controller;
 pub mod radar_lens;
+pub mod sensor_sets;
 pub mod util;
 pub use animating_helper::{TnuaAnimatingState, TnuaAnimatingStateDirective};
 pub use basis_action_traits::{